+++ mopack/builders/custom.py
@@ -93,7 +93,7 @@ class CustomBuilder(DirectoryBuilder):
     def clean(self, metadata, pkg):
         if self.outdir:
             path_values = pkg.path_values(metadata)
-            shutil.rmtree(path_values['builddir'], ignore_errors=True)
+            shutil.rmtree(path_values[self.outdir + 'dir'], ignore_errors=True)
 
     def build(self, metadata, pkg):
         path_values = pkg.path_values(metadata)